test: split tests into distinct groups
