+++ test/init.test.js
@@ -1,0 +1,42 @@
+import test from 'tape-promise/tape'
+import LndGrpc from '../src'
+import { remoteHost } from './helpers/grpc'
+
+const { host, cert, macaroon } = remoteHost
+const grpcOptions = { host, cert, macaroon }
+
+test('initialize', t => {
+  t.plan(14)
+  const grpc = new LndGrpc(grpcOptions)
+  t.equal(grpc.state, 'ready', 'should start in the ready state')
+  t.equal(grpc.can('activateWalletUnlocker'), true, 'can activateWalletUnlocker')
+  t.equal(grpc.can('activateLightning'), true, 'can activateLightning')
+  t.equal(grpc.can('disconnect'), false, 'can not disconnect')
+  t.equal(grpc.options, grpcOptions, 'should store constructor options on the options property')
+  t.true(grpc.services, 'should have a services property')
+  t.true(grpc.services.WalletUnlocker, `should have WalletUnlocker service`)
+  t.true(grpc.services.Lightning, `should have Lightning service`)
+  t.true(grpc.services.Autopilot, `should have Autopilot service`)
+  t.true(grpc.services.ChainNotifier, `should have ChainNotifier service`)
+  t.true(grpc.services.Invoices, `should have Invoices service`)
+  t.true(grpc.services.Router, `should have Router service`)
+  t.true(grpc.services.Signer, `should have Signer service`)
+  t.true(grpc.services.WalletKit, `should have WalletKit service`)
+})
+
+test('constructor (paths)', t => {
+  t.plan(3)
+  const grpc = new LndGrpc(grpcOptions)
+  t.equal(grpc.options.host, host, 'should extract the host')
+  t.equal(grpc.options.cert, cert, 'should extract the cert')
+  t.equal(grpc.options.macaroon, macaroon, 'should extract the macaroon')
+})
+
+test('constructor (lndconnect)', t => {
+  t.plan(3)
+  const lndconnectUri = `lndconnect://${host}?cert=${cert}&macaroon=${macaroon}`
+  const grpc = new LndGrpc({ lndconnectUri })
+  t.equal(grpc.options.host, host, 'should extract the host')
+  t.equal(grpc.options.cert, cert, 'should extract the cert')
+  t.equal(grpc.options.macaroon, macaroon, 'should extract the macaroon')
+})
